test: add node environment
