+++ packages/graphql-hooks/test-jsdom/unit/GraphQLClient.test.ts
@@ -1,8 +1,7 @@
 import { GraphQLClient } from '../../src'
 import { TextEncoder } from 'util'
 import fetchMock from 'jest-fetch-mock'
-import { FormData, File as FormDataFile } from 'formdata-node'
-import { Readable } from 'stream'
+import { FormData, File } from 'formdata-node'
 import {
   createMockCache,
   createMockResponse,
@@ -459,52 +458,6 @@ describe('GraphQLClient', () => {
 
       it('does not set Content-Type header', () => {
         expect(fetchOptions.headers).not.toHaveProperty('Content-Type')
-      })
-    })
-
-    describe('with files in Node JS', () => {
-      const originalFormData = global.FormData
-
-      const client = new GraphQLClient({ ...validConfig, FormData })
-      const blob = new Blob(['sample'], { type: 'text/plain' })
-      const file = new File([blob], 'sample.txt')
-
-      const operation = { query: '', variables: { a: file } }
-      const fetchOptions = client.getFetchOptions(operation)
-
-      beforeAll(() => {
-        //@ts-ignore
-        delete global.FormData
-      })
-
-      afterAll(() => {
-        global.FormData = originalFormData
-      })
-
-      it('sets body conforming to the graphql multipart request spec', () => {
-        const actual = fetchOptions.body as FormData
-
-        expect(actual).toBeInstanceOf(FormData)
-        expect(actual.get('operations')).toBe(
-          '{"query":"","variables":{"a":null}}'
-        )
-        expect(actual.get('map')).toBe('{"1":["variables.a"]}')
-        const actualFile = actual.get('1') as unknown as File
-        expect(actualFile).toBeInstanceOf(FormDataFile)
-        expect(actualFile.name).toBe('sample.txt')
-      })
-
-      it('does not set Content-Type header', () => {
-        expect(fetchOptions.headers).not.toHaveProperty('Content-Type')
-      })
-
-      it('throws if no FormData polyfill provided', () => {
-        const client = new GraphQLClient(validConfig)
-        const operation = { query: '', variables: { a: new Readable() } }
-
-        expect(() => client.getFetchOptions(operation)).toThrow(
-          'GraphQLClient: FormData must be polyfilled or passed in new GraphQLClient({ FormData })'
-        )
       })
     })
   })
