+++ packages/graphql-hooks/test-jsdom/unit/useManualQuery.test.ts
