+++ packages/graphql-hooks/test-node/sample.txt
@@ -1,0 +1,1 @@
+sample
